requirements.txt
grpcio==1.68.0
grpcio-tools==1.68.0
torch
transformers
PyYAML
psutil
PyQt6
loguru
